chore(webpack): Give up using webpack-serve for now
It works almost fine but a 404 error occurs on hot reloading everytime.
It requests `hot-update.json` to the wrong port address. It seems that
the hot client uses a port of the HTML page, not the one it is served from.
I couldn't find a way to solve this.

webpack-serve config:

```js
serve: {
  content: DEST_DIR,
  host: '0.0.0.0',
  port: PORT,
  hot: {
    hmr: false,
  },
  // Add CORS header.
  // https://github.com/webpack-contrib/webpack-serve/pull/63
  // https://github.com/webpack-contrib/webpack-serve#add-on-features
  add: (app, middleware) => {
    app.use((ctx, next) => {
      ctx.response.set('Access-Control-Allow-Origin', '*');
      next();
    });
    middleware.webpack();
    middleware.content();
  },
},
```