feat(user): Restore last_synced_at column
We removed this column before, but it was wrong.
At that time we considered the `last_synced_at` would be
same as the latest `updated_on_pocket_at` of the user's bookmarks.
However the `updated_on_pocket_at` is updated when
we change the bookmark's state from Wracket as well. For example
it changes if its favorite/unfavirote state is toggled.
Therefore we cannot determine when is the last synchronization
was done from the `updated_on_pocket_at` column.