chore(ci): Change trigger branch rules
Currently my production server is not running all the time.
It means we cannot push to the master branch when the server
does not run because it triggers a deployment to the stopped server.

So we change a trigger branch for deployments to allow us to
push to the master branch anytime.